# docs/swagger.yaml
definitions:
...
  dto.ReqCreateParameter:
    properties:
      code:
        maxLength: 255
        type: string
      desc:
        type: string
      name:
        maxLength: 255
        type: string
      parent_id:
        type: string
      type:
        type: string
      value:
        type: string
    required:
    - code
    - name
    type: object
...
  dto.ReqUpdateParameter:
    properties:
      code:
        maxLength: 255
        type: string
      desc:
        type: string
      name:
        maxLength: 255
        type: string
      parent_id:
        type: string
      type:
        type: string
      value:
        type: string
    required:
    - code
    - name
    type: object
...
  dto.RespParameter:
    properties:
      code:
        type: string
      created_at:
        type: string
      deletable:
        type: boolean
      desc:
        type: string
      id:
        type: string
      name:
        type: string
      parent:
        $ref: '#/definitions/dto.RespParameterParent'
      type:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  dto.RespParameterIndex:
    properties:
      code:
        type: string
      created_at:
        type: string
      deletable:
        type: boolean
      id:
        type: string
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  dto.RespParameterParent:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  response.NonPaginationResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
  response.PaginationResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: Welcome to the API documentation for the Base Template Web Application.
    This comprehensive guide is designed to help developers seamlessly integrate and
    interact with our platform's functionalities. Whether you're building new features,
    enhancing existing ones, or troubleshooting, this documentation provides all the
    necessary resources and information.
  title: Base Template API Documentation
  version: 0.0-beta
paths:
...
  /v1/parameter:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of parameters with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: Filter by parameter types (array)
        in: query
        items:
          type: string
        name: types
        type: array
      - collectionFormat: csv
        description: Filter by parameter names (array)
        in: query
        items:
          type: string
        name: names
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved parameters
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RespParameterIndex'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Get list of parameters with pagination
      tags:
      - Parameter
    post:
      consumes:
      - application/json
      description: Create a new parameter with provided code, name, value, and description
      parameters:
      - description: Parameter creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateParameter'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created parameter
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespParameter'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Create a new parameter
      tags:
      - Parameter
  /v1/parameter/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing parameter by ID
      parameters:
      - description: Parameter UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted parameter
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "400":
          description: Bad request - invalid UUID
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Parameter not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Delete parameter
      tags:
      - Parameter
    get:
      consumes:
      - application/json
      description: Retrieve a single parameter by its ID
      parameters:
      - description: Parameter UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved parameter
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespParameter'
              type: object
        "400":
          description: Bad request - invalid UUID
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Parameter not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Get parameter by ID
      tags:
      - Parameter
    put:
      consumes:
      - application/json
      description: Update an existing parameter's information
      parameters:
      - description: Parameter UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated parameter data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqUpdateParameter'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated parameter
          schema:
            allOf:
            - $ref: '#/definitions/response.NonPaginationResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RespParameter'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "404":
          description: Parameter not found
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Update parameter
      tags:
      - Parameter
  /v1/parameter/export:
    get:
      consumes:
      - application/json
      description: Export parameters to Excel file (.xlsx) with optional search and
        filter. Same search and filter logic as index but without pagination.
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: Filter by parameter types (array)
        in: query
        items:
          type: string
        name: types
        type: array
      - collectionFormat: csv
        description: Filter by parameter names (array)
        in: query
        items:
          type: string
        name: names
        type: array
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel file with parameters data
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.NonPaginationResponse'
      security:
      - BearerAuth: []
      summary: Export parameters to Excel
      tags:
      - Parameter
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token (ex: Bearer eyJhbGciOiJIU....)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"